#80: adapted Login and registration page to updated form component (& start on trying to recover lost stylings from merge)
